#23 move format methods to subfolder
